ocr-service/requirements.txt
Flask
gunicorn
pytesseract
Pillow
prometheus-flask-exporter